backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
oracledb>=2.0.0
requests>=2.31.0
cachetools>=5.3.0
pydantic>=2.7.0
python-multipart>=0.0.9
yfinance>=0.2.36
